--- a/src/test/java/pingis/services/ChallengeServiceTest.java
+++ b/src/test/java/pingis/services/ChallengeServiceTest.java
@@ -24,7 +24,7 @@ import pingis.entities.User;
 import pingis.repositories.ChallengeRepository;
 
 @RunWith(SpringRunner.class)
-@ContextConfiguration(classes = {Application.class})
+@ContextConfiguration(classes = {ChallengeService.class})
 public class ChallengeServiceTest {
 
   @Autowired