Initializing the full Spring application context in a test case is super slow. SUPER slow. See the benchmarks:

Before:
```
gradle clean unittest  2.29s user 0.16s system 6% cpu 37.908 total
gradle clean unittest  2.37s user 0.11s system 6% cpu 37.886 total
gradle clean unittest  2.32s user 0.11s system 6% cpu 36.785 total
```
Average test time: 37.526

After:
```
gradle clean unittest  2.30s user 0.17s system 10% cpu 23.907 total
gradle clean unittest  2.20s user 0.09s system 9% cpu 24.326 total
gradle clean unittest  2.20s user 0.11s system 9% cpu 23.806 total
```
Average test time: 24.013

The only thing done here was replacing Application.class in ContextConfiguration with more appropriate ones.