@laurpulk 2017-08-16
Actually I noticed this problem earlier but forgot to fix it.

Nice. Work.